\name{aldh2}
\alias{aldh2}
\title{ALDH2 markers and Alcoholism}
\description{This data set contains eight ALDH2 markers 
  and Japanese patients and controls.
}
\usage{data(aldh2)}
\format{A data frame:
\describe{
\item{id}{subject id}
\item{y}{a variable taking value 0 for controls and 1 for Schizophrenia}
\item{D12S2070.a1}{D12S2070 allele a1}
\item{D12S2070.a2}{D12S2070 allele a2}
\item{D12S839.a1}{D12S839 allele a1}
\item{D12S839.a2}{D12S839 allele a2}
\item{D12S821.a1}{D12S821 allele a1}
\item{D12S821.a2}{D12S821 allele a2}
\item{D12S1344.a1}{D12S1344 allele a1}
\item{D12S1344.a2}{D12S1344 allele a2}
\item{EXON12.a1}{EXON12 allele a1}
\item{EXON12.a2}{EXON12 allele a2}
\item{EXON1.a1}{EXON1 allele a1}
\item{EXON1.a2}{EXON1 allele a2}
\item{D12S2263.a1}{D12S2263 allele a1}
\item{D12S2263.a2}{D12S2263 allele a2}
\item{D12S1341.a1}{D12S1341 allele a1}
\item{D12S1341.a2}{D12S1341 allele a2}
}
  
There are genotypes for 8 loci, with a prefix name
(e.g., "EXON12") and a suffix for each of two alleles (".a1" and ".a2").

The eight markers loci follows the following map (base pairs)

\tabular{lr}{
D12S2070   \tab (> 450 000),\cr
D12S839    \tab (> 450 000),\cr
D12S821    \tab (\eqn{\sim}{~} 400 000),\cr
D12S1344   \tab (   83 853),\cr
EXON12     \tab (    0),\cr
EXON1      \tab (   37 335),\cr
D12S2263   \tab (   38 927),\cr
D12S1341   \tab (> 450 000)
}
}
\source{
  Prof Ian Craig of Oxford and SGDP Centre, KCL
}
\keyword{datasets}
