% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{allele.recode}
\alias{allele.recode}
\title{Allele recoding}
\usage{
allele.recode(a1, a2, miss.val = NA)
}
\arguments{
\item{a1}{first allele.}

\item{a2}{second allele.}

\item{miss.val}{missing value.}
}
\description{
Allele recoding
}
