% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_hmm.R
\name{altraphase}
\alias{altraphase}
\title{HMM for genotyping and phasing for alleotetraploid individual}
\usage{
altraphase(
  datafile = NULL,
  alignment = NULL,
  res_file = NULL,
  formula = mode ~ 1 | read_pos + qua + hap_nuc + qua:hap_nuc,
  max_iter = 20,
  seed = 0,
  tol = 1e-05,
  ncores = 2,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{datafile}{Input data file path (Generated from calling `call_aln`)}

\item{alignment}{Universal reference genome (Generated from calling `call_aln`)}

\item{res_file}{Output the results in a file. Default: NULL}

\item{formula}{Formula pass to `mnlogit`}

\item{max_iter}{Iteration in HMM. Default:20}

\item{seed}{Set seed. Default:0}

\item{tol}{Convergence tolerance. Default:1e-05}

\item{ncores}{Cores to use when calling `mnlogit`}

\item{verbose}{Print process. Default: FALSE}

\item{...}{Set options}
}
\value{
A list contain the phased haplotypes, SNPs and SNP locations and other auxiliary values.
}
\description{
HMM for genotyping and phasing for alleotetraploid individual
}
\examples{
datFile <- system.file("extdata", "out.txt", package = "GAP")
refFile <- system.file("extdata", "uni.fa", package = "GAP")
res <- altraphase(datafile = datFile, alignment = refFile, max_iter = 10)
}
