\name{apoeapoc}
\alias{apoeapoc}
\title{APOE/APOC1 markers and Schizophrenia}
\description{This data set contains APOE/APOC1 markers 
  and Chinese Schizophrenic patients and controls.
}
\usage{data(apoeapoc)}
\format{A data frame:
\describe{
\item{id}{subject id}
\item{y}{a variable taking value 0 for controls and 1 for Schizophrenia}
\item{sex}{sex}
\item{age}{age}
\item{apoe.a1}{APOE allele a1}
\item{apoe.a2}{APOE allele a2}
\item{apoc.a1}{APOC allele a1}
\item{apoc.a2}{APOC allele a2}
}
  
The last six variables are age, sex and genotypes for APOE and APOC 
with suffixes for each of two alleles (".a1" and ".a2").
}
\source{
  Dr JJ Shi of Western China Medical University
}
\keyword{datasets}
