% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asplot.R
\name{asplot}
\alias{asplot}
\title{Regional association plot}
\usage{
asplot(
  locus,
  map,
  genes,
  flanking = 1000,
  best.pval = NULL,
  sf = c(4, 4),
  logpmax = 10,
  pch = 21
)
}
\arguments{
\item{locus}{Data frame with columns c("CHR", "POS", "NAME", "PVAL", "RSQR") containing association results.}

\item{map}{Genetic map, i.e, c("POS","THETA","DIST").}

\item{genes}{Gene annotation with columns c("START", "STOP", "STRAND", "GENE").}

\item{flanking}{Flanking length.}

\item{best.pval}{Best p value for the locus of interest.}

\item{sf}{scale factors for p values and recombination rates, smaller values are necessary for gene dense regions.}

\item{logpmax}{Maximum value for -log10(p).}

\item{pch}{Plotting character for the SNPs to be highlighted, e.g., 21 and 23 refer to circle and diamond.}
}
\description{
Regional association plot
}
\details{
This function obtains regional association plot for a particular locus, based on
the information about recombinatino rates, linkage disequilibria between the
SNP of interest and neighbouring ones, and single-point association tests p values.

Note that the best p value is not necessarily within locus in the original design.
}
\examples{
\dontrun{
require(gap.datasets)
asplot(CDKNlocus, CDKNmap, CDKNgenes)
title("CDKN2A/CDKN2B Region")
asplot(CDKNlocus, CDKNmap, CDKNgenes, best.pval=5.4e-8, sf=c(3,6))
}

}
\references{
\insertRef{saxena07}{gap}
}
\author{
{Paul de Bakker, Jing Hua Zhao, Shengxu Li}
}
\keyword{hplot}
