% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{call_aln}
\alias{call_aln}
\title{Prepare data for `GAP`}
\usage{
call_aln(
  ref_nameA = NULL,
  ref_nameB = NULL,
  ref_fsa = NULL,
  ref_sam = NULL,
  alnA = NULL,
  alnB = NULL,
  out_file = NULL,
  uni_geno_file = NULL
)
}
\arguments{
\item{ref_nameA}{Reference name of genome A which shows in file *refSam*.
With  ':' to separate genome name and base index and '-' to separate the begin (0-based)
and end index(begin + length of reference A, 1-based)}

\item{ref_nameB}{Reference name of genome B which shows in file *refSam*.
Similar way to define as genome A.}

\item{ref_fsa}{Fasta file contains alignment of A and B genome ('-' stands for deletion).}

\item{ref_sam}{Sam file align B genome to A genome}

\item{alnA}{Sam file align reads to A genome}

\item{alnB}{Sam file align reads to B genome}

\item{out_file}{File path and name of the outputted file}

\item{uni_geno_file}{File path and name of the universal fasta genome}
}
\value{
Write the input datafile and universal reference genome file for `altraphase`
}
\description{
Prepare data for `GAP`
}
\note{
We could use MUMmer(and MUMmer3.20) to get sam file (ref.sam) and fasta file (ref.fa):
(sequence names in A.fa and B.fa should be named with the begin (0-based) and end (1-based)
index relative to the whole genomes)
nucmer -p ch1 --mum A.fa B.fa
delta-filter -1 ch1.delta > ch1_1to1.delta
MUMmer3.20/delta2maf ch1_1to1.delta > ch1_1to1.maf
cat ch1_1to1.maf | awk '/^s/{print ">" $2  "\\n" $7}' > ref.fa
nucmer --sam-long=refsam --mum A.fa B.fa
bioawk -c fastx '{  print "@SQ " "SN:" $name, "LN:" length($seq) }' < A.fa > header
cat header refsam > ref.sam
}
\examples{
alignment <- system.file("extdata", "ref.fa", package = "GAP")
ref_sam <- system.file("extdata", "ref.sam", package = "GAP")
alnA <- system.file("extdata", "aln0A.sam", package = "GAP")
alnB <- system.file("extdata", "aln0B.sam", package = "GAP")
out_file <- system.file("extdata", "out.txt", package = "GAP")
uni_geno_file <- system.file("extdata", "uni.fa", package = "GAP")
call_aln(ref_nameA = "Genome_A:0-1000", ref_nameB = "Genome_B:0-1000",
ref_fsa = alignment, ref_sam = ref_sam, alnA = alnA, alnB = alnB,
out_file = out_file, uni_geno_file = uni_geno_file)
}
