\name{CBS_breakpoints_for_GAPconstruction}
\alias{CBS_breakpoints_for_GAPconstruction}
\title{
CBS_breakpoints_for_GAPconstruction
}
\description{

}
\usage{

}
\arguments{

}
\value{

}
\note{

}
\author{
Tatiana Popova\cr
\email{tatiana.popova@curie.fr}\cr
\url{http://bioinfo.curie.fr/projects/snp_gap/}
}
\seealso{

}
\examples{

}
\keyword{}