% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpid.R
\name{chr_pos_a1_a2}
\alias{chr_pos_a1_a2}
\title{SNP id by chr:pos+a1/a2}
\usage{
chr_pos_a1_a2(
  chr,
  pos,
  a1,
  a2,
  prefix = "chr",
  seps = c(":", "_", "_"),
  uppercase = TRUE
)
}
\arguments{
\item{chr}{Chromosome.}

\item{pos}{Position.}

\item{a1}{Allele 1.}

\item{a2}{Allele 2.}

\item{prefix}{Prefix of the identifier.}

\item{seps}{Delimiters.}

\item{uppercase}{A flag to return in upper case.}
}
\value{
Identifier.
}
\description{
SNP id by chr:pos+a1/a2
}
\details{
This function generates unique identifiers for variants
}
\examples{
# rs12075
chr_pos_a1_a2(1,159175354,"A","G",prefix="chr",seps=c(":","_","_"),uppercase=TRUE)
}
