% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circos.cnvplot.R
\name{circos.cnvplot}
\alias{circos.cnvplot}
\title{circos plot of CNVs.}
\usage{
circos.cnvplot(data)
}
\arguments{
\item{data}{CNV data containing chromosome, start, end and freq.}
}
\value{
None.
}
\description{
The function plots frequency of CNVs.
}
\examples{
circos.cnvplot(cnv)
}
