% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circos.mhtplot.R
\name{circos.mhtplot}
\alias{circos.mhtplot}
\title{circos Manhattan plot with gene annotation}
\usage{
circos.mhtplot(data, glist)
}
\arguments{
\item{data}{Data to be used.}

\item{glist}{A gene list.}
}
\value{
None.
}
\description{
The function generates circos Manhattan plot with gene annotation.
}
