\name{CopyNumber_and_Genotype}
\alias{CopyNumber_and_Genotype}
\title{
CopyNumber_and_Genotype
}
\description{

}
\usage{

}
\arguments{

}
\value{

}
\note{

}
\author{
Tatiana Popova\cr
\email{tatiana.popova@curie.fr}\cr
\url{http://bioinfo.curie.fr/projects/snp_gap/}
}
\seealso{

}
\examples{

}
\keyword{}