% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_qtls.R
\name{find_qtls}
\alias{find_qtls}
\title{Distance-based signal identification}
\usage{
find_qtls(
  d,
  Chromosome = "Chromosome",
  Position = "Position",
  MarkerName = "MarkerName",
  Allele1 = "Allele1",
  Allele2 = "Allele2",
  EAF = "Freq1",
  Effect = "Effect",
  StdErr = "StdErr",
  log10P = "log10P",
  N = "N",
  radius = 1e+06
)
}
\arguments{
\item{d}{input data.}

\item{Chromosome}{chromosome.}

\item{Position}{position.}

\item{MarkerName}{RSid or SNPid.}

\item{Allele1}{effect allele.}

\item{Allele2}{other allele.}

\item{EAF}{effect allele frequency.}

\item{Effect}{b.}

\item{StdErr}{SE.}

\item{log10P}{-log(P).}

\item{N}{sample size.}

\item{radius}{a flanking distance.}
}
\value{
The function lists QTLs and meta-information.
}
\description{
Distance-based signal identification
}
\details{
This function implements an iterative merging algorithm to identify signals.
The setup follows output from METAL.
}
\examples{
\dontrun{
  f <- "ZPI_dr.p.gz"
  varlist=c("Chromosome","Position","MarkerName","Allele1","Allele2",
            "Freq1","FreqSE","MinFreq","MaxFreq",
            "Effect","StdErr","log10P","Direction",
            "HetISq","HetChiSq","HetDf","logHetP","N")
  d <- read.table(f,col.names=varlist,check.names=FALSE)
  find_qtls(d)
}
}
