\name{fsnps}
\alias{fsnps}
\title{A case-control data involving four SNPs with missing genotype}
\description{
This is a simulated data
}
\usage{data(hla)}
\format{A data frame
\describe{
\item{id}{subject id}
\item{y}{a column of 0/1 indicating case/control}
\item{site1.a1}{SNP 1 allele a1}
\item{site1.a2}{SNP 1 allele a2}
\item{site2.a1}{SNP 2 allele a1}
\item{site2.a2}{SNP 2 allele a2}
\item{site3.a1}{SNP 3 allele a1}
\item{site3.a2}{SNP 3 allele a2}
\item{site4.a1}{SNP 4 allele a1}
\item{site4.a2}{SNP 4 allele a2}
}
  
The last eight variables are genotypes for 4 SNPs, coded in characters
}
\source{
Dr Sebastien Lissarrague of Genset
}
\keyword{datasets}
