\name{package:GAP}
\docType{package}
\alias{GAP}
\title{
	Genome Alteration Print (GAP)
}
\description{

}
\details{
	\tabular{ll}{
		Package: \tab GAP\cr
		Type: \tab Package\cr
		Version: \tab 0.0-1\cr
		Date: \tab 2013-09-14\cr
		License: \tab GPL (>= 2)\cr
		LazyLoad: \tab yes\cr
	}
}
\author{
	\tabular{ll}{
		author(s): \tab Tatiana Popova\cr
				   \tab tatiana.popova@curie.fr\cr
				   \tab \url{http://bioinfo.curie.fr/projects/snp_gap/}\cr\cr
				   \tab \cr
		maintainer(s): \tab David Brown\cr
					   \tab david.brown@bordet.be\cr
					   \tab \url{http://homepages.ulb.ac.be/~dbrown/}\cr\cr
	}
}
\references{
	T.Popova, E. Manie, D. Stoppa-Lyonnet, G. Rigaill, E. Barillot and M.H. Stern.
	Genome Alteration Print (GAP): a tool to visualize and mine complex cancer
	genomic profiles obtained by SNP arrays. Genome Biology, 10:R128, 2009.\cr
	
}
\note{
	\url{http://genomebiology.com/2009/10/11/R128}
}
\keyword{package}