% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gap.function.final.R
\name{GAP_bayesian_p2_logl}
\alias{GAP_bayesian_p2_logl}
\title{GAP Bayesian Log-likelihood for State 2 (Trait 1 Only)}
\usage{
GAP_bayesian_p2_logl(x0, beta, sd, rho, alpha, t1, t2, r)
}
\arguments{
\item{x0}{A numeric value representing the effect size for trait 2 (beta_12)}

\item{beta}{A numeric vector of length 3 containing the observed effect sizes
(beta_01, beta_12, beta_cc)}

\item{sd}{A numeric vector of length 3 containing the standard errors}

\item{rho}{A numeric value for the correlation parameter}

\item{alpha}{A numeric value for the alpha parameter (not used in this function)}

\item{t1}{A numeric value for the first prior standard deviation parameter}

\item{t2}{A numeric value for the second prior standard deviation parameter}

\item{r}{A numeric value for the prior correlation parameter (not used in this function)}
}
\value{
A numeric value representing the negative log-likelihood
}
\description{
Calculates the log-likelihood for the GAP model state where there is
association with trait 1 only (beta_01 != 0, beta_12 = 0).
}
\details{
This is an internal function used in the optimization process
for the GAP model. It calculates the likelihood for the state where
beta_01 is free to vary while beta_12 is constrained to be zero.
}
\seealso{
\code{\link{GAP_bayesian_lrt}}
}
