% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gap.function.final.R
\name{GAP_bayesian_p4_logl_0_1_null}
\alias{GAP_bayesian_p4_logl_0_1_null}
\title{GAP Bayesian Log-likelihood for State 4 with Trait 1 Null}
\usage{
GAP_bayesian_p4_logl_0_1_null(x0, beta, sd, rho, alpha, t1, t2, r)
}
\arguments{
\item{x0}{A numeric value representing the effect size for trait 2 (beta_12)}

\item{beta}{A numeric vector of length 3 containing the observed effect sizes
(beta_01, beta_12, beta_cc)}

\item{sd}{A numeric vector of length 3 containing the standard errors}

\item{rho}{A numeric value for the correlation parameter}

\item{alpha}{A numeric value for the alpha parameter (not used in this function)}

\item{t1}{A numeric value for the first prior standard deviation parameter}

\item{t2}{A numeric value for the second prior standard deviation parameter}

\item{r}{A numeric value for the prior correlation parameter}
}
\value{
A numeric value representing the negative log-likelihood
}
\description{
Calculates the log-likelihood for the GAP model state 4 with the constraint
that beta_01 = 0 (trait 1 null hypothesis).
}
\details{
This is an internal function used in the likelihood ratio test
for the GAP model. It calculates the likelihood for the pleiotropic state
with the constraint that beta_01 = 0.
}
\seealso{
\code{\link{GAP_bayesian_lrt}}
}
