% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gap.function.final.R
\name{GAP_prior_00_integral_v2}
\alias{GAP_prior_00_integral_v2}
\title{GAP Prior Integral for State 00 (No Association)}
\usage{
GAP_prior_00_integral_v2(beta, sd, rho, t1, t2, r, alpha)
}
\arguments{
\item{beta}{A matrix of effect sizes with columns for beta_01, beta_12, beta_cc}

\item{sd}{A matrix of standard errors corresponding to beta}

\item{rho}{A numeric value for the correlation parameter}

\item{t1}{A numeric value for the first prior standard deviation parameter}

\item{t2}{A numeric value for the second prior standard deviation parameter}

\item{r}{A numeric value for the prior correlation parameter}

\item{alpha}{A numeric value for the alpha parameter (not used in this state)}
}
\value{
A numeric vector of log-likelihood values for each SNP
}
\description{
Calculates the log-likelihood integral for the GAP model state 00,
representing no association with either trait.
}
\details{
This function calculates the likelihood for the null state where
there is no association with either trait. It uses a multivariate normal
distribution for the first two traits and a univariate normal for the third.
}
\seealso{
\code{\link{GAP_prior_v2}}
}
