% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gap.function.final.R
\name{GAP_prior_cc_integral_v2}
\alias{GAP_prior_cc_integral_v2}
\title{GAP Prior Integral for State cc (Pleiotropic Association)}
\usage{
GAP_prior_cc_integral_v2(beta, sd, rho, t1, t2, r, alpha)
}
\arguments{
\item{beta}{A matrix of effect sizes with columns for beta_01, beta_12, beta_cc}

\item{sd}{A matrix of standard errors corresponding to beta}

\item{rho}{A numeric value for the correlation parameter}

\item{t1}{A numeric value for the first prior standard deviation parameter}

\item{t2}{A numeric value for the second prior standard deviation parameter}

\item{r}{A numeric value for the prior correlation parameter}

\item{alpha}{A numeric value for the alpha parameter (affects the variance scaling)}
}
\value{
A numeric vector of log-likelihood values for each SNP
}
\description{
Calculates the log-likelihood integral for the GAP model state cc,
representing pleiotropic association with both traits (beta_01 != 0, beta_12 != 0).
}
\details{
This function calculates the likelihood for the pleiotropic state where
there is association with both traits. The alpha parameter affects the variance
scaling and correlation structure for the third trait (beta_cc).
}
\seealso{
\code{\link{GAP_prior_v2}}
}
