% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gap.function.final.R
\name{GAP_prior_v2}
\alias{GAP_prior_v2}
\title{GAP Prior Likelihood Function (Version 2)}
\usage{
GAP_prior_v2(x0, beta_m, se_m, alpha, rho)
}
\arguments{
\item{x0}{A numeric vector of length 6 containing the parameters to optimize:
\itemize{
  \item x0[1]: s1 parameter (standard deviation component 1)
  \item x0[2]: s2 parameter (standard deviation component 2)
  \item x0[3]: s3 parameter (correlation component)
  \item x0[4]: log(p1) - log probability of state 1
  \item x0[5]: log(p2) - log probability of state 2
  \item x0[6]: log(p3) - log probability of state 3
}}

\item{beta_m}{A matrix of effect sizes with columns for beta_01, beta_12, beta_cc}

\item{se_m}{A matrix of standard errors corresponding to beta_m}

\item{alpha}{A numeric value for the alpha parameter in the GAP model}

\item{rho}{A numeric value for the correlation parameter}
}
\value{
A numeric value representing the negative log-likelihood
}
\description{
Calculates the negative log-likelihood for the GAP (Genetic Association with
Pleiotropy) prior model. This is an internal function used in the optimization
process to estimate prior parameters.
}
\details{
This function implements the GAP prior model which considers four
possible states for genetic associations:
\enumerate{
  \item State 00: No association with either trait
  \item State 01: Association with trait 1 only
  \item State 12: Association with trait 2 only
  \item State cc: Association with both traits (pleiotropy)
}

The function calculates the likelihood for each state and combines them
using the prior probabilities to obtain the overall likelihood.
}
\seealso{
\code{\link{GAP_prior_00_integral_v2}}, \code{\link{GAP_prior_01_integral_v2}},
  \code{\link{GAP_prior_12_integral_v2}}, \code{\link{GAP_prior_cc_integral_v2}}
}
