% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_gap_theme.R
\docType{data}
\name{gap_theme}
\alias{gap_theme}
\title{thematic golem theme}
\format{
An object of class \code{bs_theme_with_preset} (inherits from \code{bs_version_5}, \code{bs_theme}, \code{sass_bundle}) of length 1.
}
\usage{
gap_theme
}
\value{
bslib theme
}
\description{
thematic golem theme
}
\keyword{datasets}
