% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gc.lambda.R
\name{gc.lambda}
\alias{gc.lambda}
\title{Estionmation of the genomic control inflation statistic (lambda)}
\usage{
gc.lambda(p, logscale = FALSE)
}
\arguments{
\item{p}{A vector of p values.}

\item{logscale}{A logical variable for \code{log.p}}
}
\value{
Estimate of inflation factor.
}
\description{
Estionmation of the genomic control inflation statistic (lambda)
}
\examples{
set.seed(12345)
p <- runif(100)
gc.lambda(p)
lp <- -log10(p)
gc.lambda(lp,logscale=TRUE)
}
