% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_b_se.R
\name{get_b_se}
\alias{get_b_se}
\title{Get b and se from AF, n, and z}
\usage{
get_b_se(f, n, z)
}
\arguments{
\item{f}{Allele frequency.}

\item{n}{Sample size.}

\item{z}{z-statistics.}
}
\value{
b and se.
}
\description{
The function obtains effect size and its standard error.
}
\examples{
\dontrun{
  get_b_se(0.6396966,23991,4.7245)
#               b          se
# [1,] 0.04490488 0.009504684
# eQTLGen
  cis_pQTL <- merge(read.delim('eQTLGen.lz') \%>\%
              filter(GeneSymbol=="LTBR"),read.delim("eQTLGen.AF"),by="SNP") \%>\%
              mutate(data.frame(get_b_se(AlleleB_all,NrSamples,Zscore)))
}
}
