% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrs.R
\name{grid2d}
\alias{grid2d}
\title{Two-dimensional grid}
\usage{
grid2d(chrlen, plot = TRUE, cex = 0.6)
}
\arguments{
\item{chrlen}{Lengths of chromosomes; e.g., hg18, hg19 or hg38.}

\item{plot}{A flag for plot.}

\item{cex}{A scaling factor for labels.}
}
\description{
This function build 2-d grids
}
