% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsmr.R
\name{gsmr}
\alias{gsmr}
\title{Mendelian randomization analysis}
\usage{
gsmr(data, X, Y, alpha = 0.05, other_plots = FALSE)
}
\arguments{
\item{data}{Data to be used.}

\item{X}{Exposure.}

\item{Y}{Outcome.}

\item{alpha}{type I error rate for confidence intervals.}

\item{other_plots}{To add funnel and forest plots.}
}
\value{
The result and plots.
}
\description{
The function initially intends to rework on GSMR outputs, but it would be appropriate for general use.
}
\examples{
library(cowplot)
library(ggplot2)
library(gap)
r <- gsmr(mr, "LIF.R", c("CAD","FEV1"))
}
