% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2G.R
\name{h2G}
\alias{h2G}
\title{Heritability and its variance}
\usage{
h2G(V, VCOV, verbose = TRUE)
}
\arguments{
\item{V}{Variance estimates.}

\item{VCOV}{Variance-covariance matrix.}

\item{verbose}{Detailed output.}
}
\value{
A list of phenotypic variance/heritability estimates and their variances.
}
\description{
Heritability and its variance
}
