% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2GE.R
\name{h2GE}
\alias{h2GE}
\title{Heritability and its variance when there is an environment component}
\usage{
h2GE(V, VCOV, verbose = TRUE)
}
\arguments{
\item{V}{Variance estimates.}

\item{VCOV}{Variance-covariance matrix.}

\item{verbose}{Detailed output.}
}
\value{
A list of phenotypic variance/heritability/GxE interaction esimates and their variances.
}
\description{
Heritability and its variance when there is an environment component
}
