% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2l.R
\name{h2l}
\alias{h2l}
\title{Heritability under the liability threshold model}
\usage{
h2l(K = 0.05, P = 0.5, h2, se, verbose = TRUE)
}
\arguments{
\item{K}{Disease prevalence.}

\item{P}{Phenotypeic variance.}

\item{h2}{Heritability estimate.}

\item{se}{Standard error.}

\item{verbose}{Detailed output.}
}
\value{
A list of the input heritability estimate/standard error and their counterpart under liability threshold model, the normal deviate..
}
\description{
Heritability under the liability threshold model
}
