% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{hg18}
\alias{hg18}
\title{hg18 information}
\format{
A vector containing lengths of chromosomes.
}
\usage{
hg18
}
\description{
GRCh36 information
}
\details{
generated from GRCh.R.
}
\keyword{datasets}
\keyword{internal}
