% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{hg19}
\alias{hg19}
\title{hg19 information}
\format{
A vector containing lengths of chromosomes.
}
\usage{
hg19
}
\description{
GRCh37 information
}
\details{
generated from GRCh.R.
}
\keyword{datasets}
\keyword{internal}
