% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{hg38}
\alias{hg38}
\title{hg38 information}
\format{
A vector containing lengths of chromosomes.
}
\usage{
hg38
}
\description{
GRCh38 information
}
\details{
generated from GRCh.R.
}
\keyword{datasets}
\keyword{internal}
