\name{hla}
\alias{hla}
\title{HLA markers and Schizophrenia}
\description{This data set contains HLA markers DRB, DQA, DQB
  and 271 Schizophrenia patients and controls.
}
\usage{data(hla)}
\format{A data frame containing 271 rows and 8 columns:
\describe{
\item{id}{subject id}
\item{y}{a variable taking value 0 for controls and 1 for Schizophrenia}
\item{DQR.a1}{DQR allele a1}
\item{DQR.a2}{DQR allele a2}
\item{DQA.a1}{DQA allele a1}
\item{DQA.a2}{DQA allele a2}
\item{DQB.a1}{DQB allele a1}
\item{DQB.a2}{DQB allele a2}
}
  
The last six variables are genotypes for 3 HLA loci, with a prefix
name (e.g., "DQB") and a suffix for each of two alleles (".a1" and
".a2").
}
\source{
Dr Padraig Wright of Pfitzer
}
\keyword{datasets}
