% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_utils_server.R
\name{if_x_NULL_y}
\alias{if_x_NULL_y}
\alias{\%||\%}
\title{\%\|\|\%}
\usage{
x \%||\% y
}
\arguments{
\item{x, y}{Two elements to test, one potentially `NULL`}
}
\description{
If x is `NULL`, return y, otherwise return x
}
\examples{
NULL \%||\% 1
}
