\name{inf1}
\docType{data}
\alias{inf1}
\title{A data containing protein panel}
\description{This data is used to illustrate cis/trans classification, containing the following columns:
\preformatted{
}
                                                        Target Target.Short
1                                        Osteoprotegerin (OPG)          OPG
2                            C-X-C motif chemokine 11 (CXCL11)       CXCL11
3                     TNF-related activation cytokine (TRANCE)       TRANCE
4                                               Axin-1 (AXIN1)        AXIN1
5                               C-C motif chemokine 25 (CCL25)        CCL25
6 Tumor necrosis factor (Ligand) superfamily member 12 (TWEAK)        TWEAK
  UniProtID      Gene chrom     Start       End
1    O00300 TNFRSF11B     8 119935796 119964439
2    O14625    CXCL11     4  76954835  76962568
3    O14788   TNFSF11    13  43136872  43182149
4    O15169     AXIN1    16    337440    402673
5    O15444     CCL25    19   8117651   8127534
6    O43508   TNFSF12    17   7452208   7464925
}
\usage{data(inf1)}
\format{A data frame containing 92 rows and 7 columns}
\source{Undisclosed}
  
\keyword{datasets}
