% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpid.R
\name{inv_chr_pos_a1_a2}
\alias{inv_chr_pos_a1_a2}
\title{Retrieval of chr:pos+a1/a2 according to SNP id}
\usage{
inv_chr_pos_a1_a2(chr_pos_a1_a2, prefix = "chr", seps = c(":", "_", "_"))
}
\arguments{
\item{chr_pos_a1_a2}{SNP id.}

\item{prefix}{Prefix of the identifier.}

\item{seps}{Delimiters of fields.}
}
\value{
A data.frame with the following variables.
\describe{
\item{chr}{Chromosome.}
\item{pos}{Position.}
\item{a1}{Allele 1.}
\item{a2}{Allele 2.}
}
}
\description{
This function obtains information embedded in  unique identifiers.
}
\examples{
# rs12075
inv_chr_pos_a1_a2("chr1:159175354_A_G",prefix="chr",seps=c(":","_","_"))
}
