% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invnormal.R
\name{invnormal}
\alias{invnormal}
\title{Inverse normal transformation}
\usage{
invnormal(x)
}
\arguments{
\item{x}{Data with missing values.}
}
\value{
Transformed value.
}
\description{
Inverse normal transformation
}
\examples{
x <- 1:10
z <- invnormal(x)
plot(z,x,type="b")
}
