% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrs.R
\name{ixy}
\alias{ixy}
\title{Conversion of chrosome name from strings}
\usage{
ixy(x)
}
\arguments{
\item{x}{Chromosome name in strings}
}
\value{
As indicated.
}
\description{
This function converts 1:22, X, Y back to 1:24.
}
