% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KCC.R
\name{KCC}
\alias{KCC}
\title{Disease prevalences in cases and controls}
\usage{
KCC(model, GRR, p1, K)
}
\arguments{
\item{model}{disease model (one of "multiplicative","additive","recessive","dominant","overdominant").}

\item{GRR}{genotype relative risk.}

\item{p1}{disease allele frequency.}

\item{K}{disease prevalence in the whole population.}
}
\value{
A list of two elements:
\itemize{
\item pprime prevlence in cases.
\item p prevalence in controls.
}
}
\description{
Disease prevalences in cases and controls
}
\details{
KCC calculates disease prevalences in cases and controls for a given genotype relative risk,
allele frequency and prevalencen of the disease in the whole population. It is used by tscc
and pbsize2.
}
