% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log10p.R
\name{log10p}
\alias{log10p}
\title{log10(p) for a normal deviate z}
\usage{
log10p(z)
}
\arguments{
\item{z}{normal deviate.}
}
\value{
log10(P)
}
\description{
log10(p) for a normal deviate z
}
\examples{
log10p(100)
}
\author{
James Peters
}
