% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log10pvalue.R
\name{log10pvalue}
\alias{log10pvalue}
\title{log10(p) for a P value including its scientific format}
\usage{
log10pvalue(p = NULL, base = NULL, exponent = NULL)
}
\arguments{
\item{p}{value.}

\item{base}{base part in scientific format.}

\item{exponent}{exponent part in scientific format.}
}
\value{
log10(P)
}
\description{
log10(p) for a P value including its scientific format
}
\examples{
log10pvalue(1e-323)
log10pvalue(base=1,exponent=-323)
}
