\name{mhtdata}
\docType{data}
\alias{mhtdata}
\title{An example data for Manhattan plot}
\description{
This example contains p values for a list of SNPs whose information 
regarding chromosome, position and reference seqeuence as with gene
annotation is obtained separately.}
\usage{data(mhtdata)}
\format{A data frame}
\source{Dr Tuomas Kilpelainen at the MRC Epidemiology Unit}
\references{
None
}
\keyword{datasets}
