% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhtplot2.R
\name{mhtplot2}
\alias{mhtplot2}
\title{Manhattan plot with annotations}
\usage{
mhtplot2(data, control = mht.control(), hcontrol = hmht.control(), ...)
}
\arguments{
\item{data}{a data frame with three columns representing chromosome, position and p values.}

\item{control}{A list produced by \code{\link[=mht.control]{mht.control()}} controlling plot behaviour.}

\item{hcontrol}{A list produced by \code{\link[=hmht.control]{hmht.control()}} defining highlighted markers or regions and labels.}

\item{...}{Additional graphical arguments passed to \code{\link[graphics:plot]{graphics::plot()}}
(e.g. \code{pch}, \code{bg}, \code{xlab}, \code{ylab}, \code{ylim}).
If \code{mar} is supplied, the default margins are not modified.}
}
\value{
shown on or saved to the appropriate device.
}
\description{
Manhattan plot with annotations
}
\details{
To generate Manhattan plot with annotations. The function is generic and for instance could be used for genomewide
p values or any random variable that is uniformly distributed. By default, a log10-transformation is applied.
Note that with real chromosomal positions, it is also appropriate to plot and some but not all chromosomes.

It is possible to specify options such as xlab, ylim and font family when the plot is requested for data in other
context. The example uses only chromosomes 14 and 20 of \insertCite{denhoed13}{gap}.
}
\examples{
\dontrun{
mdata <- within(hr1420,{
  c1<-colour==1
  c2<-colour==2
  c3<-colour==3
  colour[c1] <- 62
  colour[c2] <- 73
  colour[c3] <- 552
})
mdata <- mdata[,c("CHR","POS","P","gene","colour")]
ops <- mht.control(colors=rep(c("lightgray","gray"),11),yline=1.5,xline=2)
hops <- hmht.control(data=subset(mdata,!is.na(gene)),boxed=TRUE)
v <- "Verdana"
ifelse(Sys.info()['sysname']=="Windows", windowsFonts(ffamily=windowsFont(v)),
       ffamily <- v)
tiff("mh.tiff", width=.03937*189, height=.03937*189/2, units="in", res=1200,
     compress="lzw")
par(las=2, xpd=TRUE, cex.axis=1.8, cex=0.4)
mhtplot2(with(mdata,cbind(CHR,POS,P,colour)),ops,hops,pch=19,
         ylab=expression(paste(plain("-"),log[10],plain("p-value"),sep=" ")),
         family="ffamily")
axis(2,pos=2,at=seq(0,25,5),family="ffamily",cex=0.5,cex.axis=1.1)
dev.off()

# To exemplify the use of chr, pos and p without gene annotation
# in response to query from Vallejo, Roger <Roger.Vallejo@ARS.USDA.GOV>
opar <- par()
par(cex=0.4)
ops <- mht.control(colors=rep(c("lightgray","lightblue"),11),yline=2.5,xline=2)
mhtplot2(data.frame(mhtdata[,c("chr","pos","p")],gene=NA,color=NA),ops,xlab="",ylab="")
axis(2,at=1:16)
title("data in mhtplot used by mhtplot2")
par(opar)
}
}
\references{
\insertRef{denhoed13}{gap}
}
\author{
Jing Hua Zhao
}
