% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhtplot3d.R
\name{mhtplot3d}
\alias{mhtplot3d}
\title{3D Manhattan plot}
\usage{
mhtplot3d(d, chrlen = gap::hg19, zmax = 300)
}
\arguments{
\item{d}{Data in mhtplot2d() format.}

\item{chrlen}{Lengths of chromosomes for specific build: hg18, hg19, hg38.}

\item{zmax}{Maximum -log10p to truncate, above which they would be set to this value.}
}
\value{
A plotly figure.
}
\description{
3D Manhattan plot
}
\examples{

\dontrun{
INF <- Sys.getenv("INF")
d <- read.csv(file.path(INF,"work","INF1.merge.cis.vs.trans"),as.is=TRUE)
r <- mhtplot3d(d,zmax=300)
htmlwidgets::saveWidget(r,file=file.path(INF,"INF1.latest.html"))
r
}
}
