\name{nep499}
\docType{data}
\alias{nep499}
\title{A study of Alzheimer's disease with eight SNPs and APOE}
\description{
This is a study of the neprilysin gene and sporadic Alzheimer's disease
in Chinese. There are 257 cases and 242 controls, each with eight SNPs
detecting through denaturing high-performance liquid chromatography
(DHPLC).
}
\usage{data(nep499)}
\format{A data frame}
\source{
J Shi, S Zhang, M Tang, C Ma, J Zhao, T Li, X Liu, Y Sun, Y Guo, H Han, Y Ma, Z Zhao.
Mutation Screening and Association Study of the Neprilysin Gene in Sporadic Alzheimer's Disease in Chinese Persons.
J Gerontol A: Bio Sci Med Sci 60:301-306, 2005}
\keyword{datasets}
