% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedtodot_verbatim.R
\name{pedtodot_verbatim}
\alias{pedtodot_verbatim}
\title{Pedigree-drawing with graphviz}
\usage{
pedtodot_verbatim(f)
}
\arguments{
\item{f}{A data.frame containing pedigrees, each with pedigree id, individual id, father id, mother id, sex and affection status.}
}
\value{
For each pedigree, a Graphviz DOT file is written to the working directory.
}
\description{
Pedigree-drawing with graphviz
}
\details{
Read a GAS or LINKAGE pedigree and write a Graphviz DOT representation for each pedigree.

This is a verbatim translation of the original \code{pedtodot} implemented in Bash/awk.
To check independently, try \code{xclip < pedtodot_verbatim.R} and paste into an R session.
}
\note{
Adapted from Bash/awk script by David Duffy
}
\examples{
\dontrun{
# pedigree p3 in pedtodot
  pedtodot_verbatim(p3)
}
}
