% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pqtl2dplotly.R
\name{pqtl2dplotly}
\alias{pqtl2dplotly}
\title{2D pQTL plotly}
\usage{
pqtl2dplotly(d, chrlen = gap::hg19)
}
\arguments{
\item{d}{Data in pqtl2dplot() format.}

\item{chrlen}{Lengths of chromosomes for specific build: hg18, hg19, hg38.}
}
\value{
A plotly figure.
}
\description{
2D pQTL plotly
}
\examples{

\dontrun{
INF <- Sys.getenv("INF")
d <- read.csv(file.path(INF,"work","INF1.merge.cis.vs.trans"),as.is=TRUE)
r <- pqtl2dplotly(d)
htmlwidgets::saveWidget(r,file=file.path(INF,"INF1.pqtl2dplotly.html"))
r
}
}
