% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvalue.R
\name{pvalue}
\alias{pvalue}
\title{P value for a normal deviate}
\usage{
pvalue(z, decimals = 2)
}
\arguments{
\item{z}{normal deviate.}

\item{decimals}{number of decimal places.}
}
\value{
P value as a string variable.
}
\description{
P value for a normal deviate
}
\examples{
pvalue(-1.96)
}
