% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ReadGRMBin}
\alias{ReadGRMBin}
\title{A function to read GRM binary files}
\usage{
ReadGRMBin(prefix, AllN = FALSE, size = 4)
}
\arguments{
\item{prefix}{file root.}

\item{AllN}{a logical variable.}

\item{size}{size.}
}
\description{
A function to read GRM binary files
}
\details{
Modified from GCTA documentation
}
