% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revStrand.R
\name{revStrand}
\alias{revStrand}
\title{Allele on the reverse strand}
\usage{
revStrand(allele)
}
\arguments{
\item{allele}{Allele to reverse.}
}
\value{
Allele on the reverse strand.
}
\description{
Allele on the reverse strand
}
\details{
The function obtains allele on the reverse strand.
}
\examples{
\dontrun{
  alleles <- c("a","c","G","t")
  reverse_strand(alleles)
}
}
