% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runshinygap.R
\name{runshinygap}
\alias{runshinygap}
\title{Start shinygap}
\usage{
runshinygap(...)
}
\arguments{
\item{...}{Additional arguments passed to the 'runApp' function from the 'shiny' package.}
}
\value{
These are design specific.
}
\description{
Start shinygap
}
\details{
This function starts the interactive 'shinygap' shiny web application that allows for flexible model specification.

The 'shiny' based web application allows for flexible model specification for the implemented study designs.
}
