% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_utils_server.R
\name{rv}
\alias{rv}
\title{Typing reactiveValues is too long}
\usage{
rv(...)
}
\arguments{
\item{...}{Objects that will be added to the reactivevalues object. All of
these objects must be named.}
}
\description{
Typing reactiveValues is too long
}
\keyword{internal}
