% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_scatter_display.R
\name{scatter_plot}
\alias{scatter_plot}
\title{scatter plot utility function}
\usage{
scatter_plot(df, x_var, y_var, col_var, alpha_var, size_var)
}
\arguments{
\item{df}{\emph{(data.frame)} The dataset containing the variables to plot.}

\item{x_var}{\emph{(character)} Name of the variable for the x-axis.}

\item{y_var}{\emph{(character)} Name of the variable for the y-axis.}

\item{col_var}{\emph{(character)} Name of the variable for the color aesthetic.}

\item{alpha_var}{\emph{(numeric)} Transparency level of points (0 to 1).}

\item{size_var}{\emph{(numeric)} Size of points.}
}
\value{
The return value, if any, from executing the utility.
}
\description{
A utils function
}
