\name{snca}
\docType{data}
\alias{snca}
\title{A study of Parkinson's disease and SNCA makers}
\description{
A study of Parkinson's disease and controls with markers of SNCA region. m770, int4
and int3 are SNPs, SNCA is dinucleotide repeat. The column abc indicates if a subject 
is familial Parkinson's (+), sporadic (-), or controls (CON). Races involved are
American Indians (AI), African American (B), and the rest are Caucasians. Diagnosis
also included possible (POS), probable (PRO) and definite PDs. AON is age onset, or
unknown (UN).
}

\usage{data(snca)}
\format{A data frame}
\source{Prof Abbas Parsian of University of Arkansas Medical Sciences}
\references{Parsian et al. ASHG 2005, Toronto}

\keyword{datasets}
