\name{summaryPlots_perChrom}
\alias{summaryPlots_perChrom}
\title{
summaryPlots_perChrom
}
\description{

}
\usage{

}
\arguments{

}
\value{

}
\note{

}
\author{
Tatiana Popova\cr
\email{tatiana.popova@curie.fr}\cr
\url{http://bioinfo.curie.fr/projects/snp_gap/}
}
\seealso{

}
\examples{

}
\keyword{}