% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{WriteGRM}
\alias{WriteGRM}
\title{A function to write GRM file}
\usage{
WriteGRM(prefix = 51, id, N, GRM)
}
\arguments{
\item{prefix}{file root.}

\item{id}{id.}

\item{N}{sample size.}

\item{GRM}{a GRM.}
}
\description{
A function to write GRM file
}
