% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{WriteGRMBin}
\alias{WriteGRMBin}
\title{A function to write GRM binary file}
\usage{
WriteGRMBin(prefix, grm, N, id, size = 4)
}
\arguments{
\item{prefix}{file root.}

\item{grm}{a GRM.}

\item{N}{Sample size.}

\item{id}{id.}

\item{size}{size.}
}
\description{
A function to write GRM binary file
}
