% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrs.R
\name{xy}
\alias{xy}
\title{Conversion of chromosome names to strings}
\usage{
xy(x)
}
\arguments{
\item{x}{(alpha)numeric value indicating chromosome.}
}
\value{
As indicated.
}
\description{
Conversion of chromosome names to strings
}
\details{
This function converts x=1:24 to 1:22, X, Y
}
