\name{apoeapoc}
\docType{data}
\alias{apoeapoc}
\title{APOE/APOC1 markers and schizophrenia}
\description{This data set contains APOE/APOC1 markers 
  and Chinese Schizophrenic patients and controls. Variable id is subject
id and y takes value 0 for controls and 2 for Schizophrenia.

The last six variables are age, sex and genotypes for APOE and APOC 
with suffixes for each of two alleles (".a1" and ".a2").

}
\usage{data(apoeapoc)}
\format{A data frame}
\source{Dr JJ Shi of Western China Medical University}
  
\keyword{datasets}
