% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ci2bse}
\alias{ci2bse}
\title{Effect size and standard error from confience interval and P value}
\usage{
ci2bse(ci, p, or = TRUE, alpha = 0.05)
}
\arguments{
\item{ci}{confidence interval.}

\item{p}{P value.}

\item{or}{a flag indicating the confidence interval is based on OR.}

\item{alpha}{confidence level.}
}
\value{
Based on confidence interval of OR, the function provides a list containing estimates
\itemize{
\item b effect size (log(OR))
\item se standard error
\item direction \code{sign} of b.
}
}
\description{
Effect size and standard error from confience interval and P value
}
\details{
The confidence interval actually contains aliased information so we
\itemize{
\item use the upper limit only.
\item add -/0/+ for sign(b).
}
}
\examples{
# rs3784099 and breast cancer recurrence/mortality (PMID22232737)
bse <- ci2bse("1.28-1.72",1e-7)
print(bse)
# Vectors of CIs and Ps
ci2 <- c("1.28-1.72","1.25-1.64")
p2 <- c(1e-7,3e-7)
bse <- ci2bse(ci2,p2)
print(bse)
}
