% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ci2bse}
\alias{ci2bse}
\title{Effect size and standard error from confidence interval}
\usage{
ci2bse(ci, or = TRUE, alpha = 0.05)
}
\arguments{
\item{ci}{confidence interval (CI). The delimiter between lower and upper limit is either a hyphen (-) or en dash (\enc{–}{-}).}

\item{or}{a flag indicating the confidence interval is based on odds ratio (OR).}

\item{alpha}{Type 1 error.}
}
\value{
Based on CI, the function provides a list containing estimates
\itemize{
\item b effect size (log(OR))
\item se standard error
\item direction a decrease/increase (-/+).
}
}
\description{
Effect size and standard error from confidence interval
}
\details{
Let \eqn{z \sim N(0,1)}{z ~ N(0,1)} with cutoff point \eqn{z_\alpha}{z_alpha}, then the CI is defined by confidence limits L, U as follows,
\deqn{
\begin{aligned}
L & = b - z_\alpha se \cr
U & = b + z_\alpha se
\end{aligned}
}{L = b - z_alpha se, U = b + z_alpha se}
\eqn{\Rightarrow}{==>} \eqn{U + L = 2 b}, \eqn{U - L=2 z_\alpha se}{U - L = 2 z_alpha se}. Consequently,
\deqn{
\begin{aligned}
b & = (U + L)/2 \cr
se & = (U - L)/2/z_\alpha
\end{aligned}
}{b = (U + L)/2, se = (U - L)/2/z_alpha}
for OR, \eqn{L\equiv\log(L)}{L ==> log(L)}, \eqn{U\equiv\log(U)}{U ==> log(U)}.
Additionally, \code{sign(b)}=-1, 0, 1, is labelled "-", "0", "+", respectively as in PhenoScanner.
}
\examples{
# rs3784099 and breast cancer recurrence/mortality
bse <- ci2bse("1.28-1.72")
print(bse)
# Vector input
ci2 <- c("1.28-1.72","1.25-1.64")
bse <- ci2bse(ci2)
print(bse)
}
