% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circos.cis.vs.trans.plot.R
\name{circos.cis.vs.trans.plot}
\alias{circos.cis.vs.trans.plot}
\title{circos plot of cis/trans classification}
\usage{
circos.cis.vs.trans.plot(hits, panel, id, radius = 1e+06)
}
\arguments{
\item{hits}{A text file as input data with varibles named "CHR","BP","SNP","prot".}

\item{panel}{Protein panel with prot(ein), uniprot (id) and "chr","start","end","gene".}

\item{id}{Identifier.}

\item{radius}{The flanking distance as cis.}
}
\value{
None.
}
\description{
The function implements a circos plot at the early stage of SCALLOP-INF meta-analysis.
}
\examples{
\dontrun{
  circos.cis.vs.trans.plot(hits="INF1.clumped", panel=inf1, id="uniprot")
}
}
