% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circos.mhtplot.R
\name{circos.mhtplot}
\alias{circos.mhtplot}
\title{circos Manhattan plot with gene annotation}
\usage{
circos.mhtplot(data, glist)
}
\arguments{
\item{data}{Data to be used.}

\item{glist}{A gene list.}
}
\value{
None.
}
\description{
circos Manhattan plot with gene annotation
}
\details{
The function generates circos Manhattan plot with gene annotation.
}
\examples{
\dontrun{
require(gap.datasets)
glist <- c("IRS1","SPRY2","FTO","GRIK3","SNED1","HTR1A","MARCH3","WISP3",
           "PPP1R3B","RP1L1","FDFT1","SLC39A14","GFRA1","MC4R")
circos.mhtplot(mhtdata,glist)
}

}
