% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cis.vs.trans.classification}
\alias{cis.vs.trans.classification}
\title{A cis/trans classifier}
\usage{
cis.vs.trans.classification(hits, panel, id, radius = 1e+06)
}
\arguments{
\item{hits}{Data to be used.}

\item{panel}{Panel data.}

\item{id}{Identifier.}

\item{radius}{The flanking distance for variants.}
}
\value{
The cis/trans classification.
}
\description{
The function classifies variants into cis/trans category according to a panel which contains id, chr, start, end, gene variables.
}
