\name{gap-internal}
\alias{a2g}
\alias{g2a}
\alias{g2a.c}
\alias{gc.control}
\alias{gcode}
\alias{grec2g}
\alias{hap.control}
\alias{hap.score.glm}
\alias{hap.score.podds}
\alias{is.miss}
\alias{KCC}
\alias{m2plem}
\alias{micombine}
\alias{plem2m}
\alias{revhap}
\alias{revhap.i}
\alias{toETDT}
\alias{ungcode}
\title{internal functions for gap}
\description{internal gap functions}
\details{These functions are not so frequently called by users}
\keyword{internal}
