\name{gap-internal}
\alias{Ginv}
\alias{a2g}
\alias{allele.recode}
\alias{compositeLD}
\alias{g2a}
\alias{g2a.c}
\alias{geno.recode}
\alias{gcode}
\alias{grec2g}
\alias{hap.score.glm}
\alias{hap.score.podds}
\alias{is.miss}
\alias{printBanner}
\alias{print.compositeLD}
\alias{revhap}
\alias{revhap.i}
\alias{ungcode}
\title{internal functions for gap}
\description{internal gap functions}
\details{These functions are not explicitly called by users}
\keyword{internal}
