% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrs.R
\name{grid2d}
\alias{grid2d}
\title{Two-dimensional grid}
\usage{
grid2d(
  chrlen,
  plot = TRUE,
  cex.labels = 0.6,
  xlab = "QTL position",
  ylab = "Gene position"
)
}
\arguments{
\item{chrlen}{Lengths of chromosomes; e.g., hg18, hg19 or hg38.}

\item{plot}{A flag for plot.}

\item{cex.labels}{A scaling factor for labels.}

\item{xlab}{X-axis title.}

\item{ylab}{Y-axis title.}
}
\value{
A list with two variables:
\itemize{
\item n Number of chromosomes.
\item CM Cumulative lengths starting from 0.
}
}
\description{
This function build 2-d grids
}
