% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{hg18}
\alias{hg18}
\title{Chromosomal lengths for build 36}
\format{
A vector containing lengths of chromosomes.
}
\usage{
hg18
}
\description{
Data are used in other functions.
}
\details{
generated from GRCh.R.
}
\keyword{datasets}
