% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{hg19}
\alias{hg19}
\title{Chromosomal lengths for build 37}
\format{
A vector containing lengths of chromosomes.
}
\usage{
hg19
}
\description{
Data are used in other functions.
}
\keyword{datasets}
