% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{hg38}
\alias{hg38}
\title{Chromosomal lengths for build 38}
\format{
A vector containing lengths of chromosomes.
}
\usage{
hg38
}
\description{
Data are used in other functions.
}
\keyword{datasets}
