% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhtplot.R
\name{hmht.control}
\alias{hmht.control}
\title{Controls for highlights}
\usage{
hmht.control(
  data = NULL,
  colors = NULL,
  yoffset = 0.25,
  cex = 1.5,
  boxed = FALSE
)
}
\arguments{
\item{data}{Data.}

\item{colors}{Colors.}

\item{yoffset}{Y offset.}

\item{cex}{Scaling factor.}

\item{boxed}{Label in box.}
}
\value{
A list as above.
}
\description{
Specification of highlights
}
