% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KCC.R
\name{KCC}
\alias{KCC}
\title{Disease prevalences in cases and controls}
\usage{
KCC(model, GRR, p1, K)
}
\arguments{
\item{model}{Disease model (one of "multiplicative","additive","recessive","dominant","overdominant").}

\item{GRR}{genotype relative risk.}

\item{p1}{disease allele frequency.}

\item{K}{disease prevalence in the whole population.}
}
\value{
A list of two elements,
\describe{
\item{pprime}{prevlence in cases.}
\item{p}{prevalence in controls.}
}

# 6-6-2018 MRC-Epid JHZ
}
\description{
Disease prevalences in cases and controls
}
\details{
It is used by both tscc() and pbsize2().
}
