% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl3dplotly.R
\name{qtl3dplotly}
\alias{qtl3dplotly}
\title{3D QTL plot}
\usage{
qtl3dplotly(
  d,
  chrlen = gap::hg19,
  zmax = 300,
  qtl.id = "SNPid:",
  qtl.prefix = "QTL:",
  qtl.target = "Protein:",
  qtl.gene = "target (gene):",
  xlab = "QTL position",
  ylab = "Gene position"
)
}
\arguments{
\item{d}{Data in qtl2d() format.}

\item{chrlen}{Lengths of chromosomes for specific build: hg18, hg19, hg38.}

\item{zmax}{Maximum -log10p to truncate, above which they would be set to this value.}

\item{qtl.id}{QTL id.}

\item{qtl.prefix}{QTL prefix.}

\item{qtl.target}{QTL target.}

\item{qtl.gene}{QTL gene.}

\item{xlab}{X-axis title.}

\item{ylab}{Y-axis title.}
}
\value{
A plotly figure.
}
\description{
3D QTL plot
}
\examples{
\dontrun{
INF <- Sys.getenv("INF")
d <- read.csv(file.path(INF,"work","INF1.merge.cis.vs.trans"),as.is=TRUE)
r <- qtl3dplotly(d,zmax=300)
htmlwidgets::saveWidget(r,file=file.path(INF,"INF1.qtl3dplotly.html"))
r
}
}
