% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whscore.R
\name{whscore}
\alias{whscore}
\title{Whittemore-Halpern scores for allele-sharing}
\usage{
whscore(allele, type)
}
\arguments{
\item{allele}{a matrix of alleles of affected pedigree members.}

\item{type}{0 = pairs, 1 = all.}
}
\value{
The returned value is the value of score statistic.
}
\description{
Whittemore-Halpern scores for allele-sharing
}
\details{
Allele sharing score statistics.
}
\note{
adapted from GENEHUNTER.
}
\examples{
\dontrun{
c<-matrix(c(1,1,1,2,2,2),ncol=2)
whscore(c,type=1)
whscore(c,type=2)
}

}
\references{
\insertRef{kruglyak96}{gap}

\insertRef{wh94a}{gap}

\insertRef{wh94b}{gap}
}
\author{
Leonid Kruglyak, Jing Hua Zhao
}
\keyword{utilities}
